Add wcslib-dev to Linux install instructions
The wcslib-dev package is required for install on linux, and is listed in the linux install script, but is not included in the 'building the program' instructions in the readme file. This fix adds that package to the list for people (like me) who went direct to the build instructions.